Move QgsCodeEditorWidget to a new file
