+++ python/gui/auto_generated/codeeditors/qgscodeeditor.sip.in
@@ -603,74 +603,6 @@ QFlags<QgsCodeEditor::Flag> operator|(QgsCodeEditor::Flag f1, QFlags<QgsCodeEdit
 
 
 
-class QgsCodeEditorWidget : QgsPanelWidget
-{
-%Docstring(signature="appended")
-A widget which wraps a QgsCodeEditor in additional functionality.
-
-This widget wraps an existing QgsCodeEditor object in a widget which provides
-additional standard functionality, such as search/replace tools. The caller
-must create an unparented QgsCodeEditor object (or a subclass of QgsCodeEditor)
-first, and then construct a :py:class:`QgsCodeEditorWidget` passing this object to the
-constructor.
-
-.. note::
-
-   may not be available in Python bindings, depending on platform support
-
-.. versionadded:: 3.38
-%End
-
-%TypeHeaderCode
-#include "qgscodeeditor.h"
-%End
-  public:
-
-    QgsCodeEditorWidget( QgsCodeEditor *editor /Transfer/, QWidget *parent /TransferThis/ = 0 );
-%Docstring
-Constructor for QgsCodeEditorWidget, wrapping the specified ``editor`` widget.
-
-Ownership of ``editor`` will be transferred to this widget.
-%End
-
-    QgsCodeEditor *editor();
-%Docstring
-Returns the wrapped code editor.
-%End
-
-  public slots:
-
-    void showSearchBar();
-%Docstring
-Shows the search bar.
-
-.. seealso:: :py:func:`hideSearchBar`
-
-.. seealso:: :py:func:`setSearchBarVisible`
-%End
-
-    void hideSearchBar();
-%Docstring
-Hides the search bar.
-
-.. seealso:: :py:func:`showSearchBar`
-
-.. seealso:: :py:func:`setSearchBarVisible`
-%End
-
-    void setSearchBarVisible( bool visible );
-%Docstring
-Sets whether the search bar is ``visible``.
-
-.. seealso:: :py:func:`showSearchBar`
-
-.. seealso:: :py:func:`hideSearchBar`
-%End
-
-};
-
-
-
 /************************************************************************
  * This file has been generated automatically from                      *
  *                                                                      *
