+++ python/gui/auto_generated/codeeditors/qgscodeeditorwidget.sip.in
@@ -1,0 +1,85 @@
+/************************************************************************
+ * This file has been generated automatically from                      *
+ *                                                                      *
+ * src/gui/codeeditors/qgscodeeditorwidget.h                            *
+ *                                                                      *
+ * Do not edit manually ! Edit header and run scripts/sipify.pl again   *
+ ************************************************************************/
+
+
+
+
+
+class QgsCodeEditorWidget : QgsPanelWidget
+{
+%Docstring(signature="appended")
+A widget which wraps a :py:class:`QgsCodeEditor` in additional functionality.
+
+This widget wraps an existing :py:class:`QgsCodeEditor` object in a widget which provides
+additional standard functionality, such as search/replace tools. The caller
+must create an unparented :py:class:`QgsCodeEditor` object (or a subclass of :py:class:`QgsCodeEditor`)
+first, and then construct a :py:class:`QgsCodeEditorWidget` passing this object to the
+constructor.
+
+.. note::
+
+   may not be available in Python bindings, depending on platform support
+
+.. versionadded:: 3.38
+%End
+
+%TypeHeaderCode
+#include "qgscodeeditorwidget.h"
+%End
+  public:
+
+    QgsCodeEditorWidget( QgsCodeEditor *editor /Transfer/, QWidget *parent /TransferThis/ = 0 );
+%Docstring
+Constructor for QgsCodeEditorWidget, wrapping the specified ``editor`` widget.
+
+Ownership of ``editor`` will be transferred to this widget.
+%End
+
+    QgsCodeEditor *editor();
+%Docstring
+Returns the wrapped code editor.
+%End
+
+  public slots:
+
+    void showSearchBar();
+%Docstring
+Shows the search bar.
+
+.. seealso:: :py:func:`hideSearchBar`
+
+.. seealso:: :py:func:`setSearchBarVisible`
+%End
+
+    void hideSearchBar();
+%Docstring
+Hides the search bar.
+
+.. seealso:: :py:func:`showSearchBar`
+
+.. seealso:: :py:func:`setSearchBarVisible`
+%End
+
+    void setSearchBarVisible( bool visible );
+%Docstring
+Sets whether the search bar is ``visible``.
+
+.. seealso:: :py:func:`showSearchBar`
+
+.. seealso:: :py:func:`hideSearchBar`
+%End
+
+};
+
+/************************************************************************
+ * This file has been generated automatically from                      *
+ *                                                                      *
+ * src/gui/codeeditors/qgscodeeditorwidget.h                            *
+ *                                                                      *
+ * Do not edit manually ! Edit header and run scripts/sipify.pl again   *
+ ************************************************************************/
