+++ python/gui/auto_generated/qgsqueryresultwidget.sip.in
@@ -11,6 +11,7 @@
 
 
 
+
 class QgsQueryResultWidget: QWidget
 {
 %Docstring(signature="appended")
